# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Segura",
        "description": "API RESTful para comercio electrónico con Laravel, JWT y Stripe.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
...
        "/api/payment/create-session": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Crear una sesión de pago en Stripe",
                "operationId": "c8097a553852bd3a3fb2e6d0e956279e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "product_name"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Monto a cobrar",
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "product_name": {
                                        "description": "Nombre del producto o servicio",
                                        "type": "string",
                                        "example": "Suscripción Premium"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión creada exitosamente"
                    },
                    "401": {
                        "description": "No autorizado - Token ausente o inválido"
                    },
                    "422": {
                        "description": "Error de validation"
                    },
                    "500": {
                        "description": "Error interno de Stripe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener catálogo de productos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camiseta Deportiva"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Camiseta de alta calidad para entrenamiento"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación en los datos del producto"
                    },
                    "401": {
                        "description": "Unauthenticated - Requiere Token JWT"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto específico",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "responses": {
                    "200": {
                        "description": "Producto encontrado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT con el formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}